requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
Pillow==8.3.2		# via requests
docx2pdf==0.1.7		# via requests
img2pdf==0.4.2		# via requests
PyPDF2==1.26.0		# via requests
fpdf==1.7.2		# via requests